backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
